Added a second NMS pass to the matching function
This is needed because the first NMS pass only removes duplicates in each of the screenshot grid cells, thus it is possible for duplicates to still be present after the aggregate vector is built, so adding a second NMS pass fixes that. This is still more efficient than only doing 1 pass at the end since the first pass is handled separately on different threads and the second pass that runs on the main thread doesnt have much work to do since most if not all of the duplicates are already removed.